DDF-1074 moved mbean interface out of impl package
