+++ core/catalog-core-cachecommands/src/test/java/org/codice/ddf/commands/cache/TestRemoveAllCommand.java
@@ -21,7 +21,7 @@ import static org.mockito.Mockito.verify;
 
 import org.junit.Test;
 
-import ddf.catalog.cache.solr.impl.SolrCacheMBean;
+import ddf.catalog.cache.SolrCacheMBean;
