+++ core/catalog-core-standardframework/pom.xml
@@ -201,7 +201,7 @@
                         <Embed-Transitive>true</Embed-Transitive>
                         <Private-Package>
                             ddf.catalog.cache.impl,
-                            ddf.catalog.cache.solr,
+                            ddf.catalog.cache.solr.impl,
                             ddf.catalog.event.retrievestatus,
                             ddf.catalog.impl,
                             ddf.catalog.data.impl,
@@ -216,7 +216,6 @@
                         </Private-Package>
                         <Export-Package>
                             ddf.catalog.cache,
-                            ddf.catalog.cache.solr.impl,
                             ddf.catalog.resourceretriever,
                             ddf.catalog.resource.data
                         </Export-Package>
